Add feature where params pane is intially positioned at the top of the word cloud pane

+++ CoreWordCloudView/src/au/gov/asd/tac/constellation/views/wordcloud/ui/WordCloudTopComponent.java
@@ -93,7 +93,7 @@ public final class WordCloudTopComponent extends JavaFxTopComponent<WordCloudPan
 
             // Update word cloud pane's size when window size changes
             scene.heightProperty().addListener((obv, oldVal, newVal) -> {
-                wordCloudPane.setContentHeight((int) (double) newVal);
+                wordCloudPane.setContentHeight(newVal.intValue());
             });
         });
     }
